Clean up unneccesary content
This commit removes a number of directories and files that are not
needed for the module and just clutter things up. The spec folder
was also removed, as it contains no actual spec tests. Later, a new
spec folder will be introducted with proper tests and based off of
puppetlabs spec helper.